% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accounting}
\alias{accounting}
\title{The combined accounting data sets}
\format{
A data frame with columns:
\describe{
 \item{data}{The variable is character. The data object used to calculate digit counts}
 \item{number}{The variable is numeric. The number of votes cast for the candidate}
 \item{first}{The variable is character. The first digit of number}
 \item{last}{The variable is character. The last digit of number}
}
}
\source{
\url{https://github.com/carloscinelli/benford.analysis and https://www.amazon.com/Benfords-Law-Applications-Accounting-Detection/dp/1118152859}
}
\usage{
accounting
}
\description{
The data is built to have the count in the number column with the first and last digit separated
}
\examples{
\dontrun{
accounting
}
}
\keyword{datasets}
