% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accounting_utility}
\alias{accounting_utility}
\title{A dataset of the 2010's payments data of a division of a West Coast utility company.}
\format{
A data frame with columns:
\describe{
 \item{number}{The variable is numeric. The number of votes cast for the candidate}
 \item{first}{The variable is character. The first digit of number}
 \item{last}{The variable is character. The last digit of number}
}
}
\source{
\url{https://github.com/carloscinelli/benford.analysis and https://www.amazon.com/Benfords-Law-Applications-Accounting-Detection/dp/1118152859}
}
\usage{
accounting_utility
}
\description{
The data is built to have the count in the number column with the first and last digit separated
}
\examples{
\dontrun{
accounting_utility
}
}
\keyword{datasets}
