% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benford_accounting}
\alias{benford_accounting}
\title{The counts and percentage of first digits for all data objects}
\format{
A data frame with columns:
\describe{
 \item{data}{The variable is character. The data object used to calculate digit counts}
 \item{first}{The variable is character. The first digit number}
 \item{n}{The variable is integer. The count of numbers that started with that digit}
 \item{percent}{The variable is numeric. The percent of the total for each data and country group}
 \item{benford_percent}{The variable is numeric. The expected propoprtion under Benford's law}
}
}
\source{
\url{https://github.com/carloscinelli/benford.analysis and https://www.amazon.com/Benfords-Law-Applications-Accounting-Detection/dp/1118152859}
}
\usage{
benford_accounting
}
\description{
This data has to counts by first digit for the accounting data
}
\examples{
\dontrun{
benford_accounting
}
}
\keyword{datasets}
