% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{election}
\alias{election}
\title{The election results for Iran and US presidential elections}
\format{
A data frame with columns:
\describe{
 \item{country}{The variable is character. The source country of the data}
 \item{region}{The variable is character. The region within which the election votes were tallied}
 \item{candidate}{The variable is character. The name of the electoral candidate}
 \item{number}{The variable is numeric. The number of votes cast for the candidate}
 \item{first}{The variable is character. The first digit of number}
 \item{last}{The variable is character. The last digit of number}
}
}
\source{
\url{https://github.com/midnightradio/cse140-data-programming}
}
\usage{
election
}
\description{
The data is built to have the count in the number column with the first and last digit separated
}
\examples{
\dontrun{
election
}
}
\keyword{datasets}
