% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{government_data}
\alias{government_data}
\title{A full dataset containing the card transactions for a government entity - 2010.}
\format{
A data frame with columns:
\describe{
 \item{cardnum}{The variable is character. Credit card number used for the purchase}
 \item{date}{The variable is Date. The date of the transaction}
 \item{merchnum}{The variable is character. The merchant number}
 \item{merchdescription}{The variable is character. }
 \item{merchstate}{The variable is character. The state where the merchant is located}
 \item{merchzip}{The variable is character. The zipcode of the merchant}
 \item{transtype}{The variable is character. The transaction type. A, D, P, Y}
 \item{amount}{The variable is numeric. the amount ot the transaction}
}
}
\source{
\url{https://github.com/carloscinelli/benford.analysis and https://www.amazon.com/Benfords-Law-Applications-Accounting-Detection/dp/1118152859}
}
\usage{
government_data
}
\description{
This data adds a few more variables beyond accounting_government
}
\examples{
\dontrun{
government_data
}
}
\keyword{datasets}
