% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{last_digit_accounting}
\alias{last_digit_accounting}
\title{The counts and percentage of last digits for all data objects}
\format{
A data frame with columns:
\describe{
 \item{data}{The variable is character. The data object used to calculate digit counts}
 \item{last}{The variable is character. The last digit number}
 \item{n}{The variable is integer. The count of numbers that ended with that digit}
 \item{percent}{The variable is numeric. The percent of the total for each data and country group}
 \item{last_percent}{The variable is numeric. The expected propoprtion under complete randomness}
}
}
\source{
\url{https://github.com/carloscinelli/benford.analysis and https://www.amazon.com/Benfords-Law-Applications-Accounting-Detection/dp/1118152859}
}
\usage{
last_digit_accounting
}
\description{
This data has to counts by last digit for the accounting data
}
\examples{
\dontrun{
last_digit_accounting
}
}
\keyword{datasets}
