% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pick_random}
\alias{pick_random}
\title{The counts and percentage of last digits for college students asked to pick random numbers}
\format{
A data frame with columns:
\describe{
 \item{digit}{The variable is character. The number of interest between 0-9}
 \item{n_09}{The variable is integer. The count of people that picked that digit. Note 10s were changed to 0}
 \item{percent_09}{The variable is numeric. The percentage of each digit of the total for the 0-9 digit counts}
 \item{n_last}{The variable is integer. The count of the last digit of numbers picked between 0 and 1 million.}
 \item{percent_last}{The variable is numeric. The percentage of each digt of the total for the last digit counts.}
}
}
\source{
\url{https://docs.google.com/spreadsheets/d/1TasFdyWr9xN7uWiWw0PkaFDwHYgQiC3y41YKR9CFRlA/edit#gid=0 and https://www.reddit.com/r/dataisbeautiful/comments/acow6y/asking_over_8500_students_to_pick_a_random_number/}
}
\usage{
pick_random
}
\description{
This data has to counts by last digit for the random guesses
}
\examples{
\dontrun{
pick_random
}
}
\keyword{datasets}
