% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utility_data}
\alias{utility_data}
\title{A full dataset of the 2010's payments data of a division of a West Coast utility company.}
\format{
A data frame with columns:
\describe{
 \item{vendornum}{The variable is character. Vendor Number}
 \item{date}{The variable is Date. Date of the invioce}
 \item{invnum}{The variable is character. The invoice number}
 \item{amount}{The variable is numeric. The amount on the invoice}
}
}
\source{
\url{https://github.com/carloscinelli/benford.analysis and https://www.amazon.com/Benfords-Law-Applications-Accounting-Detection/dp/1118152859}
}
\usage{
utility_data
}
\description{
This data adds a few more variables beyond accounting_utility
}
\examples{
\dontrun{
utility_data
}
}
\keyword{datasets}
