% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waitlist_finland}
\alias{waitlist_finland}
\title{The count of Finish citizens on waitlists for medical procedures}
\format{
A data frame with columns:
\describe{
 \item{country}{The variable is character. The source country of the data}
 \item{type}{The variable is character. The type of medical procedure}
 \item{details}{The variable is character. Further details about the medical procedure}
 \item{month}{The variable is character. The month of the year}
 \item{year}{The variable is numeric. Year}
 \item{number}{The variable is numeric. The number of people on the waitlist}
 \item{first}{The variable is character. The first digit of number}
 \item{last}{The variable is character. The last digit of number}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5942457/}
}
\usage{
waitlist_finland
}
\description{
The data is built to have the count in the number column with the first and last digit separated
}
\examples{
\dontrun{
waitlist_finland
}
}
\keyword{datasets}
